refactor(sec-financialfacts): extract duplicated XBRL value-parsing helpers into shared XbrlValueParser
